+++ sdk/storage/azure-storage-blob/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 12.31.0b1 (2026-07-29)
+## 12.31.0b1 (2026-07-30)
 
 ### Features Added
 - Added support for service version 2026-10-06.